Fix bug in bucketlist modification tests

+++ app/__init__.py
@@ -22,17 +22,13 @@ def create_app(configuration):
     def bucketlists():
         # Get the access token from the passed header
         auth_header = request.headers.get('Authorization')
-        print(auth_header)
         access_token = auth_header.split("Bearer ")[1]
-        print(access_token)
         if access_token:
             # Decode the token to get the user id
             user_id = User.decode_token(access_token)
-            print(user_id)
             if not isinstance(user_id, str):
                 if request.method == 'POST':
                     title = request.json.get('title')
-                    print(title)
                     if title:
                         bucketlist = Bucketlist(title=title, created_by=user_id)
                         bucketlist.save()
@@ -66,52 +62,62 @@ def create_app(configuration):
 
                     return response
             else:
-                # User_id not found, payload is anerror msg
+                # User_id not found, payload is an error msg
                 return jsonify({
-                "message": "Error, could not authenticate"
+                "message": "Error, could not authenticate. Please login first"
             }), 401
 
     @app.route('/bucketlists/<int:id>', methods=['GET', 'PUT', 'DELETE'])
     def modify_bucketlist(id, **kwargs):
-            bucketlist = Bucketlist.query.filter_by(id=id).first()
-            if not bucketlist:
-                abort(404) #Abort with a 404 not found error if buck with the id is not found
+        # Get the access token from the passed header
+        auth_header = request.headers.get('Authorization')
+        access_token = auth_header.split("Bearer ")[1]
+        if access_token:
+            # Decode the token to get the user id
+            user_id = User.decode_token(access_token)
+            if not isinstance(user_id, str):
+                bucketlist = Bucketlist.query.filter_by(id=id).first()
+                if not bucketlist:
+                    abort(404) #Abort with a 404 not found error if buck with the id is not found
 
-            if request.method == 'PUT':
-                bucketlist.title = str(request.json.get('title', ''))
-                bucketlist.save()
+                if request.method == 'PUT':
+                    bucketlist.title = request.json.get('title', '')
+                    bucketlist.save()
 
-                response = jsonify({
-                    'id': bucketlist.id,
-                    'title': bucketlist.title,
-                    'date_created': bucketlist.date_created,
-                    'date_modified': bucketlist.date_modified
-                })
-                
-                response.status_code = 200
+                    response = jsonify({
+                        'id': bucketlist.id,
+                        'title': bucketlist.title,
+                        'date_created': bucketlist.date_created,
+                        'date_modified': bucketlist.date_modified
+                    }), 200
 
-                return response
+                    return response
 
-            elif request.method == 'DELETE':
-                bucketlist.delete()
+                elif request.method == 'DELETE':
+                    bucketlist.delete()
 
-                # return make_response('Buck deleted', 200)
+                    # return make_response('Buck deleted', 200)
 
-                return jsonify({
-                    "message": "Bucketlist # {} deleted successfully".format(bucketlist.id)
-                }), 200
+                    return jsonify({
+                        "message": "Bucketlist # {} deleted successfully".format(bucketlist.id)
+                    }), 200
 
+                else:
+                    response = jsonify({
+                        'id': bucketlist.id,
+                        'title': bucketlist.title,
+                        'date_created': bucketlist.date_created,
+                        'date_modified': bucketlist.date_modified
+                    })
+
+                    response.status_code = 200
+
+                    return response
             else:
-                response = jsonify({
-                    'id': bucketlist.id,
-                    'title': bucketlist.title,
-                    'date_created': bucketlist.date_created,
-                    'date_modified': bucketlist.date_modified
-                })
-
-                response.status_code = 200
-
-                return response
+                # User_id not found, payload is an error msg
+                return jsonify({
+                "message": "Error, could not authenticate. Please login first"
+            }), 401
 
     # Import the auth blueprint and register it
     from .auth import auth_blueprint
